feat: improve the UI
